ターンを非同期で作成した

+++ src/main/java/group9/trump/controller/The_GameController.java
@@ -13,6 +13,7 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import group9.trump.model.ChamberMapper;
 import group9.trump.model.Chamber;
@@ -24,10 +25,20 @@ import group9.trump.model.The_Game_TehudaMapper;
 import group9.trump.model.The_Game_Tehuda;
 import group9.trump.model.The_Game_FieldMapper;
 import group9.trump.model.The_Game_Field;
+import group9.trump.model.TurnMapper;
+import group9.trump.model.Turn;
 import group9.trump.service.AsyncThe_Game;
+import group9.trump.service.AsyncTurn;
 
 @Controller
+@RequestMapping("/trump")
 public class The_GameController {
+
+  boolean play = false;
+  boolean saisyo = true;
+  boolean first = true;
+  int a = 6;
+
   @Autowired
   The_GameMapper TGMapper;
 
@@ -44,22 +55,30 @@ public class The_GameController {
   ChamberMapper CMapper;
 
   @Autowired
+  TurnMapper TurnMapper;
+
+  @Autowired
   AsyncThe_Game asyncThe_Game;
 
-  // ---追加
+  @Autowired
+  AsyncTurn asyncTurn;
+
   @GetMapping("/the_game")
   public String trump(Principal prin, ModelMap model) {
+    if (play == true) {
+      return "memorySkip.html";
+    }
     String loginUser = prin.getName();
     Chamber chamber = CMapper.selectByName(loginUser);
     model.addAttribute("chamber", chamber);
+    TurnMapper.delete();
     return "the_game.html";
   }
 
   @GetMapping("/the_gameMatch")
-  public String deck(Principal prin, ModelMap model) {
-    // int i = 0;
+  public String Tehuda(Principal prin, ModelMap model) {
     String loginUser = prin.getName();
-    model.addAttribute("user", loginUser);
+    model.addAttribute("loginUser", loginUser);
     int x1 = TGFMapper.selectNumber(1);
     int x2 = TGFMapper.selectNumber(2);
     int x3 = TGFMapper.selectNumber(3);
@@ -70,47 +89,40 @@ public class The_GameController {
       TGFMapper.deleteField(x1);
       x1 = TGFMapper.selectNumber(1);
     }
-    ArrayList<The_Game_Field> the_game_field1 = TGFMapper.selectAll(1);
-    // model.addAttribute("the_game_field1", the_game_field1);
     while (x2 > 1) {
       TGFMapper.deleteField(x2);
       x2 = TGFMapper.selectNumber(2);
     }
-    ArrayList<The_Game_Field> the_game_field2 = TGFMapper.selectAll(2);
-    // model.addAttribute("the_game_field2", the_game_field2);
     while (x3 < 100) {
       TGFMapper.deleteField(x3);
       x3 = TGFMapper.selectNumber(3);
     }
-    ArrayList<The_Game_Field> the_game_field3 = TGFMapper.selectAll(3);
-    // model.addAttribute("the_game_field3", the_game_field3);
     while (x4 < 100) {
       TGFMapper.deleteField(x4);
       x4 = TGFMapper.selectNumber(4);
     }
-    ArrayList<The_Game_Field> the_game_field4 = TGFMapper.selectAll(4);
-    // model.addAttribute("the_game_field4", the_game_field4);
-    // ArrayList<The_Game> tehuda = new ArrayList<>();
     ArrayList<Integer> the_game = TGMapper.selectAll();
     Collections.shuffle(the_game);
-    // model.addAttribute("the_game", the_game);
-    // while (8 > tehuda.size()) {
-    // tehuda.add(the_game.get(i));
-    // i++;
-    // }
-    // model.addAttribute("tehuda", tehuda);
-    for (int j = 0; j < the_game.size(); j++) {
+    for (int j = 0; j < 15; j++) {
       TGDMapper.insertDeck(the_game.get(j));
     }
-    // for (int k = 0; k < tehuda.size(); k++) {
-    // TGTMapper.insertTehuda(tehuda.get(k).getNumber());
-    // }
+    TurnMapper.insert(loginUser, TGTMapper.selectCount(loginUser));
+    return "the_gameMatch.html";
+  }
+
+  @GetMapping("/the_gameMatch/di")
+  public String di(Principal prin, ModelMap model) {
+    String loginUser = prin.getName();
+    model.addAttribute("loginUser", loginUser);
+    ArrayList<String> turn = TurnMapper.selectOther(loginUser);
+    model.addAttribute("turn", turn);
     return "the_gameMatch.html";
   }
 
   // ---追加
   @GetMapping("/the_gameFight")
   public String tehuda(Principal prin, ModelMap model) {
+    this.play = true;
     String loginUser = prin.getName();
     ArrayList<Integer> the_game = TGDMapper.selectAll();
     while (7 > TGTMapper.selectCount(loginUser)) {
@@ -134,30 +146,46 @@ public class The_GameController {
 
   @PostMapping("/the_gameFight/shouhi")
   @Transactional
-  public String trump2(Principal prin, ModelMap model, @RequestParam(defaultValue = "0") Integer id,
+  public String shouhi(Principal prin, ModelMap model, @RequestParam(defaultValue = "0") Integer id,
       @RequestParam(defaultValue = "0") Integer te) {
+    String error;
     String loginUser = prin.getName();
-    int a = 5;
-    int n1 = TGFMapper.selectNumber(1);
-    int n2 = TGFMapper.selectNumber(2);
-    int n3 = TGFMapper.selectNumber(3);
-    int n4 = TGFMapper.selectNumber(4);
-    int p10 = te + 10;
-    int m10 = te - 10;
-    if (id == 1 && (te > n1 || p10 == n1)) {
-      TGTMapper.deleteTehuda(te);
-      this.asyncThe_Game.syncThe_Game(1, te);
-    } else if (id == 2 && (te > n2 || p10 == n2)) {
-      TGTMapper.deleteTehuda(te);
-      this.asyncThe_Game.syncThe_Game(2, te);
-    } else if (id == 3 && (te < n3 || m10 == n3)) {
-      TGTMapper.deleteTehuda(te);
-      this.asyncThe_Game.syncThe_Game(3, te);
-    } else if (id == 4 && (te < n4 || m10 == n4)) {
-      TGTMapper.deleteTehuda(te);
-      this.asyncThe_Game.syncThe_Game(4, te);
+    if (saisyo == true) {
+      this.saisyo = false;
+      TurnMapper.update1(loginUser, TGTMapper.selectCount(loginUser));
+      ArrayList<String> others = TurnMapper.selectOther(loginUser);
+      for (int i = 0; i < TurnMapper.selectCountOther(loginUser); i++) {
+        TurnMapper.update0(others.get(i), TGTMapper.selectCount(others.get(i)));
+      }
+      this.asyncTurn.syncTurn();
+    }
+    String tasi = TurnMapper.selectUser();
+    if (tasi.equals(loginUser)) {
+      int n1 = TGFMapper.selectNumber(1);
+      int n2 = TGFMapper.selectNumber(2);
+      int n3 = TGFMapper.selectNumber(3);
+      int n4 = TGFMapper.selectNumber(4);
+      int p10 = te + 10;
+      int m10 = te - 10;
+      if (id == 1 && (te > n1 || p10 == n1)) {
+        TGTMapper.deleteTehuda(te);
+        this.asyncThe_Game.syncThe_Game(1, te);
+      } else if (id == 2 && (te > n2 || p10 == n2)) {
+        TGTMapper.deleteTehuda(te);
+        this.asyncThe_Game.syncThe_Game(2, te);
+      } else if (id == 3 && (te < n3 || m10 == n3)) {
+        TGTMapper.deleteTehuda(te);
+        this.asyncThe_Game.syncThe_Game(3, te);
+      } else if (id == 4 && (te < n4 || m10 == n4)) {
+        TGTMapper.deleteTehuda(te);
+        this.asyncThe_Game.syncThe_Game(4, te);
+      } else {
+        error = "間違え";
+        model.addAttribute("error", error);
+      }
+      TurnMapper.update1(loginUser, TGTMapper.selectCount(loginUser));
     } else {
-      String error = "間違え";
+      error = "あなたのターンではないです";
       model.addAttribute("error", error);
     }
     ArrayList<Integer> tehuda = TGTMapper.selectAll(loginUser);
@@ -173,9 +201,9 @@ public class The_GameController {
     model.addAttribute("tehuda", tehuda);
     int d = TGDMapper.selectCount();
     if (d == 0) {
-      a = 6;
+      this.a = TGTMapper.selectCount(loginUser) + 1;
     }
-    if (TGTMapper.selectCount(loginUser) <= a) {
+    if (TGTMapper.selectCount(loginUser) < this.a) {
       String end = "end";
       model.addAttribute("end", end);
     }
@@ -190,9 +218,29 @@ public class The_GameController {
     return sseEmitter;
   }
 
+  @GetMapping("/the_gameFight/deck")
+  public String kakunin(Principal prin, ModelMap model) {
+    String loginUser = prin.getName();
+    ArrayList<Integer> tehuda = TGTMapper.selectAll(loginUser);
+    int deck = TGDMapper.selectCount();
+    final ArrayList<The_Game_Field> the_game_field1 = this.asyncThe_Game.syncShowThe_Game(1);
+    final ArrayList<The_Game_Field> the_game_field2 = this.asyncThe_Game.syncShowThe_Game(2);
+    final ArrayList<The_Game_Field> the_game_field3 = this.asyncThe_Game.syncShowThe_Game(3);
+    final ArrayList<The_Game_Field> the_game_field4 = this.asyncThe_Game.syncShowThe_Game(4);
+    model.addAttribute("the_game_field1", the_game_field1);
+    model.addAttribute("the_game_field2", the_game_field2);
+    model.addAttribute("the_game_field3", the_game_field3);
+    model.addAttribute("the_game_field4", the_game_field4);
+    model.addAttribute("deck", deck);
+    model.addAttribute("tehuda", tehuda);
+    return "the_gameFight.html";
+  }
+
   @PostMapping("/the_gameFight/end")
   public String end(Principal prin, ModelMap model) {
     String loginUser = prin.getName();
+    this.first = true;
+    this.a = 6;
     final ArrayList<The_Game_Field> the_game_field1 = this.asyncThe_Game.syncShowThe_Game(1);
     final ArrayList<The_Game_Field> the_game_field2 = this.asyncThe_Game.syncShowThe_Game(2);
     final ArrayList<The_Game_Field> the_game_field3 = this.asyncThe_Game.syncShowThe_Game(3);
@@ -213,20 +261,38 @@ public class The_GameController {
     }
     ArrayList<Integer> tehuda = TGTMapper.selectAll(loginUser);
     model.addAttribute("tehuda", tehuda);
+    // -->ここasnyc TurnMapper.update0(loginUser, TGTMapper.selectCount(loginUser));
+    this.asyncTurn.syncChange(loginUser);
     return "the_gameFight.html";
   }
 
   // 追加--------------------------------------
+  @GetMapping("/the_gameFight/step2")
+  public SseEmitter step2() {
+    final SseEmitter sseEmitter = new SseEmitter();
+    this.asyncTurn.asyncShowTurn(sseEmitter);
+    return sseEmitter;
+  }
+
   @PostMapping("/the_gameFight/result")
   public String result(Principal prin, ModelMap model) {
+    String loginUser = prin.getName();
+    this.play = false;
+    this.saisyo = true;
     int d = TGDMapper.selectCount();
     int t = TGTMapper.selectCountAll();
     String win = "win";
     String lose = "lose";
+    Chamber chamber = CMapper.selectByName(loginUser);
+    asyncTurn.syncEnd();
     if (d + t < 10) {
+      CMapper.updateWin(chamber.getWin() + 1, loginUser);
+      chamber = CMapper.selectByName(loginUser);
+      model.addAttribute("chamber", chamber);
       model.addAttribute("win", win);
     } else {
       model.addAttribute("lose", lose);
+      model.addAttribute("chamber", chamber);
     }
     return "the_gameResult.html";
   }
